requirements.txt
Flask>=1.1.1
gunicorn>=19.9.0
itsdangerous>=1.1.0
Jinja2>=2.10.1
MarkupSafe==1.1.1
Werkzeug>=0.15.5
numpy>=1.9.2
scipy>=0.15.1
tensorflow-cpu>=2.5.0
scikit-learn>=0.18
